+++ optd-datafusion-repr/src/cost/base_cost/stats.rs
